requirements.txt
numpy==1.18.0
opencv-python~=4.2.0.34
Pillow~=7.2.0
tensorflow~=2.3.0
